Shell Script to format picked points to ascii #2
The Shell Script 'ascFormat.sh' formats sfipick picked points to
ascii format readable by sfdd that converts it to RSF format.

It reads points from a sfipick generated file and converts it to
RSF. If user pass 1 to the script, it cuts the first collumn of the
sfipick output and builds a RSF file with t0's coordinates. If user
pass 2 to the script, it cuts the second collumn of the sfipick output
and builds a RSF file with m0's coordinates.