Put casting parameters function in a module #9
New module inputScons.py to separate the input treatment functions.
If a key is passed through command line, example v0=2.3, the function
castingParametersFromCmd will loop over passed parameters and check
if the key passed is a parameter from parameters dictionaries used
in the functions.

If yes, the function does the casting of the parameter value and
substitute to the correct key in the parameters dictionary. The
variable ARGLIST is a SCons variable that stores all the variables
passed through command line into a python list.