Scratch of the program sfnipmodel #14
Scratch of the program sfnipmodel to receive NIP sources and
trace rays from there to acquisition surface in order to get traveltime
curves (x,t).

In the directory TDD/nipmodelDebug is an usage example of the program
for a constant velocity model, and in TDD directory is a gdb script
for set up the program debug with gdb. The program version for gdb
can be built with the following command:

~$ make nipmodel.x