Scratch of stereotomography algorithm #17
The program sfstereoniptomo uses the NIP sources to trace rays
from them to acquisition surface in order to obtain the traveltime
curve and to compare it with the CRE traveltime calculated values.
The data misfit will be used to update the velocity model.

We also add the program sfgetparameter to extract the parameters from
parameters cube for (t0,m0) location given